Start observing the Photo library for changes so that we can reload the asset grid when needed.
This is important when this library is used on iOS 14 as the user could change the limited photo library selection while the asset grid view controller is presented.
When the view is reloaded we make sure to maintain the selection while removing already selected assets that are not a part of the list anymore.